Move search field component to Search folder
