+++ src/components/Auth/form/Login.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Login = () => {
+    return <div>Login</div>;
+};
+
+export default Login;
