+++ src/components/Search/SearchResults/SingleProduct.tsx
@@ -1,8 +1,8 @@
 import { Typography } from "@mui/material";
 import Link from "next/link";
 import React, { type FC } from "react";
-import { useTypedDispatch } from "../../../../hooks/redux";
-import { setIsSearchActive } from "../../../../store/slices/searchSlice";
+import { useTypedDispatch } from "@/hooks/redux";
+import { setIsSearchActive } from "@/store/slices/searchSlice";
 
 interface Props {
     productName: string;
