+++ src/components/Search/hooks/useGetCategories.ts
