+++ src/components/Search/hooks/useMapSearchResultProducts.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import SingleProduct from "../SearchResults/SingleProduct";
-import { type IProductCard } from "../../../../types/IProductCard";
+import { type IProductCard } from "../../../types/IProductCard";
 
 export const useMapSearchResultProducts = (products: IProductCard[] | undefined) =>
     React.useMemo(() => {
